refactor: extract telegram pending edit parser
